ForMonday50
Fixed most problems as listed in Gui Polish.
Need to ask Cat how the second gui is setup. As of now the main gui can't interact with the connect channel gui, so it can't be cleared with the power or resest buttons.
Uncertain what "Default text should be set to add racer if reset or power cycled. Works for most but clear messes this up." means, but I think I fixed it.